+++ tests/testthat/test-whitepoint.R
@@ -1,11 +1,11 @@
-test_that("whitepoint_cie1931 works", {
+test_that("whitepoints_cie1931() works", {
 
   expect_error(
-    whitepoint_cie1931("foo"),
+    whitepoints_cie1931("foo"),
     "not a legal name"
   )
 
-  whitepoint <- whitepoint_cie1931("D65")
+  whitepoint <- whitepoints_cie1931("D65")
   expect_type(whitepoint, "double")
   expect_length(whitepoint, 3L)